Updated openstack/openstack Project: openstack-infra/system-config  27b0a8800caf04ca557450079f0056567e475f63
Test for sorted modules as part of linting

Merge the currently not used modulesenv target into the linters
target so that we test that modules are sorted properly which is
some kind of lint. There's no need to run an extra VM just
for checking the sorting, so include it here.

Change-Id: I547c77de92a00a42ac0841b087350144e760562f