Updated openstack/openstack Project: openstack/cinder  8c4081bfa8939cd0c229adf0cd24dfb8ceef01d0
Modify VO so that obj.get always defaults to None

Currently, CinderObjectDictCompat.get will default to None if a field
is not in the object, however if a field is present in the object
but the value is not yet set we get an error instead of None. To be
consistent with dict.get() we should default to None if the field
is not present or not set.

Change-Id: Id87efeaaeb2fb44960d8d0df9aa854dd156bff45