Updated openstack/openstack Project: openstack/oslo.db  087775381dfc1664b66c50ffd88ccce265273330
Put py34 first in the env order of tox

To solve the problem of "db type could
not be determined" on py34 we have to run first the py34 env to, then, run
py27. This patch puts py34 first on the tox.ini list of envs to avoid this
problem to happen.

Change-Id: I2365f66eee43cd36e62fe05e582c328787a0a112
Closes-bug: #1489059