Updated openstack/openstack Project: openstack/cinder  c53f24299e78f789946de13d98446dc01357d38a
Rebrand HP XP driver to now be HPE

This patch refactors the HP XP driver to be HPE XP.
This is being done because the company responsible for
this driver is now Hewlett Packard Enterprise (HPE).
The driver is now located in the cinder/volume/drivers/hpe folder.

DocImpact
Implements: blueprint rebrand-hp-xp-driver
Change-Id: Ia16d39b9fec299662c97c236215a4e1aec44171a