Updated openstack/openstack Project: openstack/nova  778d0a84394332e387871efb15c8372ccef1a906
Reduce the number of db/rpc calls to get instance rules

When getting instance rules in virt/firewall.py a for loop is used to
issue db queries for rules belonging to each individual security group
in a list of security groups that itself is fetched using a separate
query.

This can be made much more efficient by querying all rules in a single
db query joined by instance.

Change-Id: I325f9c71fecde8297842fd608ac3cfd51ea9db71
Closes-Bug: #1528041