Updated openstack/openstack Project: openstack/nova  e5b1162703718601fc4df6482c341327f7559da0
signature_utils: refactor the list of ECC curves

In the signature utils code and tests, we repeat the set of ECC curves
four times. Refactor the code so that we've only one such list.

Introduce a SignatureKeyType class which encapsulates all the info about
each signature key type and which also maintains a registry of key
types.

In a subsequent patch, this refactoring makes it easier to only register
ECC curves which are available on the platform.

Related-Bug: #1528393

Change-Id: Ic9fa1bd2149fdcdc3fc27688f6e8b141b23d8757
Signed-off-by: Mark McLoughlin <markmc@redhat.com>