Updated openstack/openstack Project: openstack-infra/project-config  9d4dd2c0659c657bf09396971cae0a12a9258cfc
Fix Zaqar gate DSVM tests to make them work

Currently all gate DSVM tests for Zaqar fail. Zaqar's gate_hook function
tries to cd to python-zaqarclient repository directory, but the
directory can't be found. Apparently python-zaqarclient repository
hasn't been cloned.

And in gate DSVM log log/devstacklog.txt.gz this line started to appear:
"The /opt/stack/new/python-zaqarclient project was not found; if this is
a gate job, add the project to the $PROJECTS variable in the job
definition."

See this patch for example of currently failing jobs:
https://review.openstack.org/#/c/259036/

Seems like the root cause of the problem is recent devstack-gate change:
https://review.openstack.org/#/c/234509

This patch adds python-zaqarclient project to the $PROJECTS variable to
Zaqar job in zaqar.yaml and Rally job in rally.yaml.

Change-Id: I9befe98fa7767c9244410f9220273c9791632f79