Updated openstack/openstack Project: openstack-infra/project-config  3d2ed91be45a67b93ee6e98b8638e8df1c3d60c2
Add new experimental Tempest job for Manila

New driver (LVM) is going to be added to Manila.
So, register job for it adding new var that contains driver name
to configure for Manila. Also, provide several more arguments for devstack
hooks, to remove dependency on job name.

Make new job run only for M+ branches in experimental pipeline
and only as nonvoting until it is merged and becomes stable.

Change-Id: I521f0419e65e27f22acbc03aaa1f0f997492e8f8