Updated openstack/openstack Project: openstack-infra/project-config  50d23009df857c030c94a40d8beeacc1494fb12a
Fix check for alphabetical sorting in layout.yaml

Previously z/tempest was in the middle of the file and needed special
treatment for sorting. Nowadays it's at the end and thus no special
treatment is needed, remove it.

With this change additions to the end of the file will now be flagged.

Change-Id: I337017b4a17e8aa086d658649613ddc61212adc8