Updated openstack/openstack Project: openstack/horizon  3f7b9abfd5a17d8f3bf4e6b60cc82634b91691be
Wrong usage of "an"

Wrong usage of "an" in the mesages:
"Width an height of the chart will be taken from chart wrapper"
"an flavor. By default this will return the flavors"
"allocate an floating IP"
"Each item is an namespace"

Should be:
"Width and height of the chart will be taken from chart wrapper"
"a flavor. By default this will return the flavors"
"allocate a floating IP"
"Each item is a namespace"

Totally 5 occasions in Horizon base code.

Change-Id: I154257da9c93b33fc7353e2e642e9e406978162b