Updated openstack/openstack Project: openstack/nova  79d6af9508b757240d3b31d726ec8719a86cd9db
signature_utils: handle ECC curve unavailability

Some ECC curves are unavailable on some platforms (like Fedora, RHEL,
and CentOS) because of legal concerns. See the bug report for more
details and history.

The cryptography backend has a elliptic_curve_supported() method
which we can use to avoid curves which are unavailable on the current
platform.

If an image signature uses one of these curves, we will return an
"Invalid signature key type" error.

Use the warnings module in the unit tests to avoid silently ignoring
this issue during testing. This warning will be captured from the
test's stderr and reported by testr.

Closes-Bug: #1528393

Change-Id: Ie25311c48b276f300fadaf1815fc4df4cb89cf8d
Signed-off-by: Mark McLoughlin <markmc@redhat.com>