Updated openstack/openstack Project: openstack-infra/project-config  2fbf3810c320e3a5ab973c2763ec61f919f6621a
Fix Tempest jobs for Manila project

Change [1] started to use comments in multiline command,
and it was done with wrong separation of comments and arguments.
So, arguments are not provided correctly to the devstack hook scripts
and they behave as nothing was provided, after their work finished we
get errors when bash tries to execute 'arguments' as separate commands.

So, fix usage of comments separating them from multiline command.

[1] I521f0419e65e27f22acbc03aaa1f0f997492e8f8

Change-Id: I5c073d38b61b21ec642ee11a28e6b23fe5662159