Updated openstack/openstack Project: openstack-infra/puppet-kibana  183dc2e05b24a2353452b2b30ba78f4bd110c76a
Remove conflicting ownership resources

/opt/kibana was being chowned to 'kibana' on every run, but
/opt/kibana/<git_version> was a vcsrepo that was being checked out as
www-data. This caused a changed file resource for every file in the git
repo for every puppet run, causing a bunch of noise in the logs and
eventually running the puppetmaster out of disk space.

Also changed directory mode to 755 because 644 doesn't make a ton of
sense.

Change-Id: I3cc511c4a5ae561ec30478d96d9745611d4b5b65