Updated openstack/openstack Project: openstack/horizon  2487c7f202e3d825e0883e25e877524881aee4be
Clean up horizon.scss - Flavor Table

A major part of the css reorg is a better organization of the css
styles themselves. horizon.scss should be a global import level file.

Launch Instance Flavor Table needs to have base Bootstrap table class.
It was noted that a bold style set for the pie charts title was
affecting the same class for the bar charts.  This was fixed.

Change-Id: If4ba59259632aa3e902129cc50c83fb9c1a36f89
Partially-Implements: blueprint horizon-theme-css-reorg
Partially-Implements: blueprint bootstrap-html-standards