Updated openstack/openstack Project: openstack/cinder  0e38f5c78a4e292808a403239c3e3fe334a2d219
Fix creating volume by snapshot for GPFS driver

For a operation of creating volume from volume snapshot,
GPFS driver is currently doing a full copy instead of clone when
it is creating a new volume. This happens when the source volume
is not in any consistency group.

With this patch, a GPFS file clone is created when both the volumes
(snapshot parent volume and new volume being created) are not in
any consistency group.

Change-Id: I0aa550842fb16b7abb8765e4aa8dce61cd0994c3
Closes-bug: #1528305