Updated openstack/openstack Project: openstack/requirements  5826647e93570b89a8d524ef93123b8876fa5924
Add voluptuous for libraries/applications data validation

The voluptuous library *can* be nicer to interact with and
is slightly more powerful (and pythonic) than the json-schema
library for data-validation.

It is already used by gnocchi and openstack-infra as an
alternative way to validate data, and tooz (a library) would
like to use it for its own data validation; so it would be
nice to let folks actually use it (especially for situations
where hacking pythonic data types onto jsonschema isn't really
a good use of time/energy).

The current version in ubuntu trusty is 0.8.2, while fedora
seems to have access to 0.8.5 (so thats why the lower bound
is being set at 0.8.2 in this change).

Is the library actively maintained? Yes

Is the library good code? Sure (for some definition of good)

Is the library python 3 compatible? Yes

Is the library license compatible? Yes

Is the library already packaged in the distros we
target (Ubuntu latest / Fedora latest)? Yes

Is the function of this library already covered by other
libraries in global-requirements.txt? Somewhat

Change-Id: Idc59dc1162bec8c64a68af5009c26472e1d2d45e