Updated openstack/openstack Project: openstack/keystone  8abf126d7e98bc8a0912034effeec04fa52fe516
Use [] where a field is required

Several places in federation.controllers used .get() where [] would
make it clear that the field is required to be present and None is
not valid and not going to work. By using [] the exception will be
raised where the problem is rather than later.

Change-Id: Ib5705662b4a406a32bb4d3d45856c566f37cf65f