Updated openstack/openstack Project: openstack-infra/subunit2sql  95ffd1f469e7045a106f88c7753f6ede5b0c049e
Add missing tests for db api functions with implicit joins

This commit adds tests for db api functions that were using implicit
joins. In a future commit the use of implicit joins will be removed
so we need to have tests to ensure the functions don't break by this
change.

Change-Id: I7ea6edf26f44a7469c11664bbfe12118aeb4882c