Updated openstack/openstack Project: openstack-infra/project-config  3d05a69c8e1b4b6ab66eef2e0925896fab9fbf88
New Project - ekko

ekko is a project that will bring block-based backup to OpenStack in a
way that is scallable by using object-storage for the backend. The
backup itself is enabled by a new feature in QEMU 2.4 that allows for
changed block tracking and incremental backups of running VMs.

A common feature I have seen requested throughout the community is
true block-based backups of VMs. I have been working on a way to solve
the scalability issue of backups for a while [1] and now that QEMU has
officially released this feature we can bring backup to OpenStack in a
way that is as scallable as OpenStack itself using Swift as a backend.

This would be different than a nova snapshot because that is not a
true backup. That does not allow for the space savings of incremental
backups nor compression or encryption which are requirements for a
good backup in my experience.

Future goals would including bringing this to other hypervisors since
Hyper-V and VMWare both support CBT and incremental backups of running
VMs. As well as potentially backing up Cinder volumes in a similiar
fashion.

[1] https://github.com/SamYaple/osdk

Change-Id: Ie0a726f0cad669593080a1910dff6ed2ae7cae71