Updated openstack/openstack Project: openstack/trove  c2d7ff4718e1aa2cde496eff396292b65dbb2baf
Fix 'cannot access' error with Redis restore

When restoring a Redis instance on Ubuntu 12.04, there has an error
in guestagent log:
"cannot access `/var/lib/redis/dump.rdb': No such file or directory"

This is because on a new instance, this file may or may not exist.
Also changed the logic around cleaning up the file before restoring the
backup in redis_impl.py as it was treating the 'restore' location as
a directory, instead of as a file.

Fixes Bug: #1528244

Change-Id: I422d68343867638c0ea6272266a21c6e73349de7