Updated openstack/openstack Project: openstack-infra/storyboard-webclient  a247659c20acbd1581fc62de3b0f9fa1850360ea
Use a modal to add projects

Adding new things (eg stories, project groups, worklists, boards) is
normally done in a pop-up modal. However, adding new projects is
inconsistent with the rest and done on a "New Project" page.

This commit adds a "New Project" modal to make the workflow consistent
with adding other types of resource.

Change-Id: I4bd5b52e662e151ce25f3e2bb27118d1ff63e57e