Updated openstack/openstack Project: openstack/zaqar  0c5a4fbe5af7a66d8f3f7b2e4b27c6e8a8ae5f31
Put py34 first in the env order of tox

To solve the problem of "db type could not be determined" on py34
we have to run first the py34 env to, then, run py27. This patch
puts py34 first on the tox.ini list of envs to avoid this problem
to happen.

Change-Id: Iabc31bfa508fbf4fad19bd0f09f389a8d3f52865
Closes-bug: #1489059