Updated openstack/openstack Project: openstack/python-novaclient  2e7e651a584d25eb808ba528ae4253277c1e8b92
Put py34 first in the env order of tox

To solve the problem of "db type could
not be determined" on py34 we have to run first the py34 env to, then, run
py27. This patch puts py34 first on the tox.ini list of envs to avoid this
problem to happen.

Change-Id: I06c6a276b75eb2a3d69e94f467f26122d856d17c
Closes-bug: #1489059