Updated openstack/openstack Project: openstack/nova  c28487c6b68a9310f8f220e2a093e0f43f0575ef
deprecate run_tests.sh

This removes the run_tests.sh script, replacing it with documentation
on how to use tox itself for testing nova things, as well as links to
learn more about both that, and how to not use tox at all if you
really care.

Why should we do this?

New people coming to the project see a file called run_tests.sh and
use it, because why would you have a file like that in the top of your
directory if that wasn't what you were supposed to use. However it's
not what we want anyone using for upstream contributions. One of the
critical differences between it and tox is that it lacks ``python
hashseed randomization``. This runs every test run with a different
forced hashseed, that exposes bugs in tests early. Test contributors
need to be running that on their tests before submission to shake out
these errors early.

In addition, because we have multiple ways in tree to run tests,
helping people debug their test failures means we have to start at a
step -1 of "how are you running tests". This adds load for helping new
folks.

What about people that want to run tests with specific system or
virtual env python setups for pre deployment testing?

Good news! Our tests comply to python standard testing interfaces, so
are runnable by any test runner (we recommend ``testr`` because that's
what tox calls under the covers). You can just run ``testr run`` at
the top level from your virtuenv or system of choice. And life is
good. The constraints we need / want for inbound tests don't apply in
these cases.

This removes a bit of confusion for new contributors, ensures that new
test contributions receive correct local testing before they are run
in our gate, and only require a very minor change in behavior for
downstream test environments.

Change-Id: I05d85a15c0db2a573abb82cb2f7dc3eeaeac0b41