Updated openstack/openstack Project: openstack/ironic  99f0bf8d3fa6ee896de2d3055815a3a3db3df707
Refactor objects into a magic registry

This adds a register hook to the objects base class that allows us to
late-import objects classes into ironic.objects. This is useful to allow
object classes to depend on other objects.

For example, this will allow us to do the Port.node_uuid munging that we
currently do in the API layer, in the objects layer instead.

This is completely taken from Nova, much thanks to comstud and dansmith
and whoever else worked on that.

Change-Id: Ic655e7af533497bda5e8e72d3bd90c27394681a7