Updated openstack/openstack Project: openstack/os-cloud-config  4809f23bbc469ebd75baa39dd5c97573ed65fed5
Remove requirement of specifying hardware stats

With introspection in place, there is no longer a hard requirement of
knowing the hardware statistics (cpu, memory, disk, arch) of nodes when
initially registering them with Ironic.

The requirement was never on the Ironic side, it was only enforced here
in os-cloud-config. Previously, to get around the requirement, one could
just specify an empty string value for these statistics in the nodes
json file.

This change makes it such that these keys do not even have to exist in
the json anymore.

Add a unit test as well that verifies this change.

Change-Id: Idcc5a9a4ac043c88fe0102626eefb327476819fd