Updated openstack/openstack Project: openstack/nova  76150fc3f15cf83839fe2bcf63c4cb649aed0457
vmware: Drop JSON for supported_instances

The virt drivers pass the value for supported_instances
to the resource tracker as a JSON encoded string. This
gets decoded as soon as it is copied to the compute node
object. The JSON is not used anymore since compute node
object was added so we will remove the encoding for all
virt drivers.

This patch removes that unnecessary encoding from the
vmware driver.

blueprint mitaka-objects

Change-Id: I0680fd236e607a9ea0909f596d31033c3600c494