Updated openstack/openstack Project: openstack/python-troveclient  bacf900c9cd03451a463e6ddbf02a189508ab8e9
Remove py26 support

As of Mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. since we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.
Closes-Bug: 1519510

Change-Id: Ifc92fdb0823a032d9b9f744494276a702a16e8b7