Updated openstack/openstack Project: openstack/nova  55d2521e1985ca6b03aa640cb0466a9b19fbd517
Unreference mocks are listed in the wrong order

There are a couple of policy checks where mocks are used to allow the
checks to complete. The mocks aren't referenced later. However they
are listed in the wrong order. While this doesn't impact this test
directly, it is confusing, and may confuse people down the road trying
to build tests based on this one.

Change-Id: Idd1da91fe8d8b1e20bef469ef67728610a9ac3c1
Closes-Bug: #1527881