Updated openstack/openstack Project: openstack-infra/project-config  f3b14c05c7dc0bbcf1d9ca8bde2c2c8f6c633f83
Skip openstack-ansible-docs gate/post jobs for juno branch

The juno branch is still active in OpenStack-Ansible, but it has no
docs in it. I've noticed that gate and post docs jobs are running
for the branch which is a waste of the node pool.

This patch does a better job of preventing the waste.

Change-Id: I7697dec6b3a4b713a8ee5c36f2ff3d936a98373f