Updated openstack/openstack Project: openstack/openstack-manuals  5ff03b6cab2fd3d4a607e9d7546fcaae368d733b
Change sg descrption typo

In the part of the doc, it describes to create security group rules not delete .
So In the decription , it should not use "deleting", it should be a typo error which copied from last paragragh.
So changed "deleting" to "checking" consistent with the above "create security group" part.

Change-Id: Ibbdc20635de82c72a0a1328c42e603bb8bb11798