Updated openstack/openstack Project: openstack/ironic-specs  bf41ab47fd3e33cc6528bc4b2a197a77a4ad2029
Don't move specs from approved directory

When specs are approved but not yet implemented, they reside in
specs/approved. After implementation, we have been moving them to
specs/<release>-implemented. Which means that any links to specs
in specs/approved are broken.

This patch leaves the specs in specs/approved, and adds a new
specs/not-implemented directory that contains soft links to all the
approved-but-not-yet-implemented specs that reside in specs/approved.
For specs that were implemented, e.g. in release 4.3, a specs/4.3
directory is created with soft links to the implemented specs (that still
reside in specs/approved).

Since we are using links to files, the specs cannot have labels
because sphinx warns of duplicate labels.

Change-Id: Icd81d68cec4bbfaa3b20e370ba1b31edb32ae1d1