Updated openstack/openstack Project: openstack-infra/reviewstats  b3d92c55c3fee2a99c2c1acffa93fb8c9f826cd0
Fix --start with existing cache

It turns out len(changes) is only the correct number of changes to
skip the first time through with an empty cache.  On subsequent
runs it skips way too far ahead in the gerrit changes and thus
never gets any new ones.

This change keeps track of the number of new changes separately
from the length of changes, which seems to make the behavior with
a cache work correctly again.

Change-Id: I1e3509d2d54b514afba64fedecc6a9f5ca408235