Updated openstack/openstack Project: openstack-infra/nodepool  761e56061c1423de7aa1a2b801057ed31850715f
Add 'public' attribute to 'networks' entry

Update the 'networks' entry to expect a different format.  Instead
of a list of dictionaries containing either 'net-id' or 'net-label,
it now expects a list of dictionaries containing at least 'name' and
optionally 'public' -- a boolean defaulting to false that indicates
that the specified network should be considered by nodepool to be the
public network.

Both 'net-id' and 'net-label' are deprecated.  This is a straightforward
transition from usage of 'net-label' to this form.  There is no
equivalent of 'net-id' in this form.

This is in service of adding Internap to the OpenStack configuration.

Change-Id: I3eb8d5183bd4c082f56853cf71842f1353977248