Updated openstack/openstack Project: openstack/nova  bdae7ff95a0f6d4c2ed721f1bece908c14a80e7e
Disable the in tree EC2 API by default

As of Kilo the in tree EC2 API is marked as deprecated and should not be
enabled by default. This will solve the following issue when working
with the default values.

---snip---
Deprecated: The in tree EC2 API is deprecated as of Kilo release and may
be removed in a future release. The stackforge ec2-api project
http://git.openstack.org/cgit/stackforge/ec2-api/ is the target
replacement for this functionality.
---snap---

Change-Id: Ia6f1d8f9c7d700498b6441a6ba37fa25b2611ef8