Updated openstack/openstack Project: openstack/nova  14eeea491601ffbaf6d2cc2ffca76341e18e72e2
xen: Drop JSON for supported_instances

The virt drivers pass the value for supported_instances
to the resource tracker as a JSON encoded string. This
gets decoded as soon as it is copied to the compute node
object. The JSON is not used anymore since compute node
object was added so we will remove the encoding for all
virt drivers.

This patch removes that unnecessary encoding from the
xen driver.

blueprint mitaka-objects

Change-Id: I0fd4e6de1f08190f8f9b500336f18030423678a1