Updated openstack/openstack Project: openstack/nova  915139a1e69d1d2e700208dc89e359cf065a8740
Drop JSON decoding for supported_instances

JSON encoding for supported_instances in the virt
drivers was a temporary measure that is not needed.
The encoding has been removed for all in tree virt
drivers. This patch removes the code that checks
for JSON encoding in the compute node object.

blueprint mitaka-objects

Change-Id: I2d0391f71436655f94f910c3b73920c6bbdd372b