Updated openstack/openstack Project: openstack/python-barbicanclient  8d3b5b3b7bcdc6d174d862c7e4ceec6e84c8fede
Fix argument order for assertEqual to (exp, obs)

assertEqual expects that the arguments provided to it
should be (expected, observed). If a particluar order
is kept as a convention, then it helps to provide a cleaner
message to the developer if Unit Tests fail.
The following patch fixes this issue.

TrivialFix

Change-Id: I6a8e11df30516f6be9a745b1f0f68ef3f2825fcf
Closes-Bug: #1259292