Updated openstack/openstack Project: openstack/requirements  443070993cfc02e54faedece5bd0b1a314734e49
add graphviz library

The project kolla needs graphviz to generate the image dependency
tree.[0]

Is the library actively maintained?
===================================
Yes. It is a active project on github[1]. And the latest release at
Nov 13, 2015[2].

Is the library good code?
=========================
It's mostly covered by unit tests and PEP8 compatible.

Is the library python 3 compatible?
===================================
Yes. It is compatible with python2.6, python2.7, python3.3, python3.4
and python3.5.

Is the library license compatible?
==================================
Yes, it's licensed under MIT.

Is the library already packaged in the distros we target?
=========================================================
No. But it is easy to package.

Is the function of this library already covered by other libraries?
===================================================================
No, there is no library interacting with Graphviz in global requirements.

[0] https://review.openstack.org/#/c/252929/
[1] https://github.com/xflr6/graphviz
[2] https://github.com/xflr6/graphviz/releases/tag/0.4.8

Change-Id: I8e07a1b69fab5f1c587470bfd2104aaba93f0050